progress: fix signed/unsigned type mismatch in pos_needs_update()
The pos parameter was declared as 'long' (signed) but the caller
passes 'size_t' (unsigned), and all related fields in
ProgressWindowData (display_pos, size, size_inc) are also size_t.

The implicit conversion from size_t to long is implementation-defined
when the value exceeds LONG_MAX. Change the parameter type to size_t
to match the calling convention and data types.

+++ progress/window.c
@@ -245,7 +245,7 @@ static bool percent_needs_update(const struct ProgressWindowData *wdata, int per
  * @param pos   Current pos
  * @retval true Progress needs an update
  */
-static bool pos_needs_update(const struct ProgressWindowData *wdata, long pos)
+static bool pos_needs_update(const struct ProgressWindowData *wdata, size_t pos)
 {
   const unsigned shift = wdata->is_bytes ? 10 : 0;
   return pos >= (wdata->display_pos + (wdata->size_inc << shift));
